Extracting a Job.Action interface
This way, the action performed by the job is independent of the job
itself. The boundaries of the data accessible to the action is clearer
(only what is available in the Params object).

This also restricts the use of reflection to building the Job.Action
object. The job object is built with a simple new, and thus can be
passed some more parameters.

Also reduced the API of Job, in part by having it handle the
rescheduling on its own.